fix js warning when running yarn start
