+++ src/component/contribute/index.js
@@ -4,10 +4,12 @@ import Layout from '../Layout';
 
 const renderContributeTitle = title => title === 'Home' ? 'Contribute' : `Contribute: ${title}`
 
-export default ({ pageContext: { content, title = 'Home' }}) => (
+const Page = ({ pageContext: { content, title = 'Home' }}) => (
   <Layout pageTitle={renderContributeTitle(title)} breadcrumbs={[{ link: '/contribute', label: 'Contribute', active: true }]} className="contribute">
     <div className="container principal">
       <ReactMarkdown source={content} />
     </div>
   </Layout>
 );
+
+export default Page
