+++ src/component/documentation/page.js
@@ -3,7 +3,7 @@ import ReactMarkdown from 'react-markdown';
 
 import Layout from '../Layout';
 
-export default ({ pageContext: { installDoc }}) => (
+const Page = ({ pageContext: { installDoc }}) => (
   <Layout pageTitle="Installation" breadcrumbs={[{ link: '/install', label: 'Install', active: true }]} className="body-green">
     <section id="install" className="about section">
       <div className="container">
@@ -12,3 +12,5 @@ export default ({ pageContext: { installDoc }}) => (
     </section>
   </Layout>
 );
+
+export default Page
