+++ src/component/install_c4/index.js
@@ -3,7 +3,7 @@ import ReactMarkdown from 'react-markdown';
 
 import Layout from '../Layout';
 
-export default ({ pageContext: { installC4Doc }}) => (
+const Page = ({ pageContext: { installC4Doc }}) => (
   <Layout section="install" pageTitle="Getting started: Class 4" breadcrumbs={[{ link: '/install/class-4', label: 'Install', active: true }]}>
     <section id="install" className="about section">
       <div className="container">
@@ -12,3 +12,5 @@ export default ({ pageContext: { installC4Doc }}) => (
     </section>
   </Layout>
 );
+
+export default Page
