+++ src/component/install_uc/index.js
@@ -3,7 +3,7 @@ import ReactMarkdown from 'react-markdown';
 
 import Layout from '../Layout';
 
-export default ({ pageContext: { installUCDoc }}) => (
+const Page = ({ pageContext: { installUCDoc }}) => (
   <Layout section="install" pageTitle="Getting started: Unified Communication" breadcrumbs={[{ link: '/uc-doc/installation/install-system', label: 'Install', active: true }]}>
     <section id="install" className="about section">
       <div className="container">
@@ -12,3 +12,5 @@ export default ({ pageContext: { installUCDoc }}) => (
     </section>
   </Layout>
 );
+
+export default Page
