+++ src/component/provisioning/external.js
@@ -3,8 +3,10 @@ import React from 'react';
 import Vendors from './components/Vendors';
 import './provisioning.scss';
 
-export default ({ pageContext: { plugins, images, content, title = 'Supported Devices' }}) => (
+const Page = ({ pageContext: { plugins, images, content, title = 'Supported Devices' }}) => (
   <div className="provisioning-vendors wazo-inc">
     <Vendors plugins={plugins} images={images} />
   </div>
 );
+
+export default Page
